Feat/descontar vagas

+++ backend/caronas/src/main/java/br/com/crescer/caronas/service/UsuarioGrupoService.java
@@ -25,12 +25,13 @@ public class UsuarioGrupoService {
     }
 
     public UsuarioGrupo save(UsuarioGrupo usuarioGrupo) {
-        if (usuarioEstaNoGrupo(usuarioGrupo)) {
+        List<UsuarioGrupo> usuariosDoGrupo = usuarioGrupoRepository.findByGrupo(usuarioGrupo.getGrupo());
+        if (usuarioEstaNoGrupo(usuarioGrupo, usuariosDoGrupo)) {
             throw new RuntimeException("O usuário já está nesse grupo");
         }
         String conteudoNotificacao = String.format("%s entrou no grupo %s", usuarioGrupo.getUsuario().getNome(), usuarioGrupo.getGrupo().getNome());
         Notificacao notificacao = new Notificacao(conteudoNotificacao, null);
-        this.enviarNotificacao(usuarioGrupo.getGrupo(), notificacao);
+        this.enviarNotificacao(usuariosDoGrupo, notificacao);
         return usuarioGrupoRepository.save(usuarioGrupo);
     }
 
@@ -39,16 +40,13 @@ public class UsuarioGrupoService {
     }
 
     public void remove(UsuarioGrupo usuarioGrupo) {
+        List<UsuarioGrupo> usuariosDoGrupo = usuarioGrupoRepository.findByGrupo(usuarioGrupo.getGrupo());
+        if (!this.usuarioEstaNoGrupo(usuarioGrupo, usuariosDoGrupo)) {
+            throw new RuntimeException("Usuário não está no grupo");
+        }
         String conteudoNotificacao = String.format("%s deixou o grupo '%s'", usuarioGrupo.getUsuario().getNome(), usuarioGrupo.getGrupo().getNome());
         Notificacao notificacao = new Notificacao(conteudoNotificacao, null);
-//        usuarioGrupo.getGrupo().getUsuarioGrupoList()
-//                .stream()
-//                .map(UsuarioGrupo::getUsuario)
-//                .forEach(usuario -> {
-//                    notificacao.setUsuario(usuario);
-//                    usuario.getNotificacaoList().add(notificacao);
-//                });
-        this.enviarNotificacao(usuarioGrupo.getGrupo(), notificacao);
+        this.enviarNotificacao(usuariosDoGrupo, notificacao);
         usuarioGrupoRepository.delete(usuarioGrupo);
     }
 
@@ -60,8 +58,8 @@ public class UsuarioGrupoService {
         return usuarioGrupoRepository.findByUsuario(usuario);
     }
 
-    public boolean usuarioEstaNoGrupo(UsuarioGrupo usuarioGrupo) {
-        return usuarioGrupo.getGrupo().getUsuarioGrupoList()
+    public boolean usuarioEstaNoGrupo(UsuarioGrupo usuarioGrupo, List<UsuarioGrupo> usuariosDoGrupo) {
+        return usuariosDoGrupo
                 .stream()
                 .map(UsuarioGrupo::getUsuario)
                 .collect(toList())
@@ -76,8 +74,8 @@ public class UsuarioGrupoService {
         return usuarioGrupoRepository.findByUsuarioAndGrupo(usuario, grupo);
     }
 
-    public void enviarNotificacao(Grupo grupo, Notificacao notificacao) {
-        grupo.getUsuarioGrupoList()
+    public void enviarNotificacao(List<UsuarioGrupo> usuariosDoGrupo, Notificacao notificacao) {
+        usuariosDoGrupo
                 .stream()
                 .map(UsuarioGrupo::getUsuario)
                 .forEach(usuario -> {
